Remove withElections HOC
